requirements.txt
groq
requests
beautifulsoup4
selenium
nltk
scikit-learn
numpy
pandas
python-dotenv
aiohttp
asyncio
lxml
urllib3
fake-useragent
textblob
spacy
transformers
torch
sentence-transformers
pypdf2
python-docx
openpyxl
chromadb
langchain
langchain-community
dateutil
re2